feat: validación de la localidad de los documentos
#300
