+++ cocemfe_sevilla/documents/templates/update_pdf.html
@@ -184,7 +184,7 @@
                                     </div>
                                     <div class="form-group col-sm-6 flex-column d-flex mb-4">
                                         <label for="id_name" class="form-control-label px-3" style="font-family: 'Arial', sans-serif; font-weight: bold;">
-                                            <i class="fas fa-map-marker-alt"></i> Ubicación<span class="text-danger"> *</span>
+                                            <i class="fas fa-map-marker-alt"></i> Localidad<span class="text-danger"> *</span>
                                         </label>
                                         {{ form.ubication }}
                                         {% for error in form.ubication.errors %}
