+++ cocemfe_sevilla/maps/tests.py
@@ -1,7 +1,6 @@
 from django.test import TestCase
 from django.test import Client
 from django.urls import reverse
-from maps.views import get_coordinates_openstreetmap, map_index, map_search
 
 class MapIndexTestCase(TestCase):
     def setUp(self):
@@ -23,13 +22,3 @@ class MapSearchTestCase(TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertTemplateUsed(response, 'maps_index.html')
 
-
-class GetCoordinatesOpenStreetMapTestCase(TestCase):
-    def test_get_coordinates_openstreetmap(self):
-        city = 'Seville'
-        latitude, longitude = get_coordinates_openstreetmap(city)
-        self.assertIsNotNone(latitude)
-        self.assertIsNotNone(longitude)
-
-        self.assertTrue(37.3886 <= latitude <= 37.4192)
-        self.assertTrue(-6.0524 <= longitude <= -5.9312)
